changes to fragment_temperature.xml

+++ app/src/main/java/com/example/currentweatherdata/models/network/WeatherResponse.kt
@@ -4,10 +4,10 @@ import com.google.gson.annotations.SerializedName
 
 data class WeatherResponse(
     @SerializedName("name") val cityName: String?,
-    @SerializedName("temp") val temp: Float? = null,
+    @SerializedName("temp") val temp: Int? = null,
     @SerializedName("description") val description: String? = null,
-    @SerializedName("temp_max") val tempMax: Float? = null, // En la imagen está primero la temperatura máxima
-    @SerializedName("temp_min") val tempMin: Float? = null,
+    @SerializedName("temp_max") val tempMax: Int? = null, // En la imagen está primero la temperatura máxima
+    @SerializedName("temp_min") val tempMin: Int? = null,
     )
 
 /*val weatherResponse: WeatherResponse? = response.body()
